BUGFIX: Odstranění přetékání měřeného času každých 65s!

+++ libraries/shutter_control/shutter_unit.cpp
@@ -39,7 +39,7 @@ void ShutterUnit::Setup()
 
 void ShutterUnit::Loop()
 {
-    unsigned int timeMs = _platform->GetTimeMs();
+    unsigned long timeMs = _platform->GetTimeMs();
 
     IoState switchIoState = _platform->ReadSwitch();
     EventId eventId = _twinButton.Loop(timeMs, switchIoState);
